+++ src/modules/auth/auth.module.ts
@@ -3,7 +3,6 @@ import { JwtModule } from '@nestjs/jwt';
 import { AuthService } from './auth.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserEntity } from 'modules/user/user.entity';
-import { AuthResolver } from './auth.resolver';
 import { JWT_SECRET } from 'core/environments';
 @Module({
   imports: [
@@ -14,7 +13,7 @@ import { JWT_SECRET } from 'core/environments';
       signOptions: { expiresIn: '30d' },
     }),
   ],
-  providers: [AuthService, AuthResolver],
+  providers: [AuthService],
   exports: [AuthService], // we will use it in every module where authentication is required
 })
 export class AuthModule {}
