+++ src/modules/auth/guards/salesman.guard.ts
@@ -1,15 +1,15 @@
 import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
 import { GqlExecutionContext } from '@nestjs/graphql';
 import { UserEntity } from 'modules/user/user.entity';
-import { AuthService } from '../auth.service';
 import { JwtService } from '@nestjs/jwt';
-import { isSalesman } from 'core/utils/authorize';
+import { isSalesman } from '../authorize';
+import { UserService } from 'modules/user/user.service';
 
 // this guard checks if requested user is a selesman
 @Injectable()
 export class SalesmanGuard implements CanActivate {
   constructor(
-    private authService: AuthService,
+    private userService: UserService,
     private jwtService: JwtService,
   ) {}
 
@@ -26,7 +26,7 @@ export class SalesmanGuard implements CanActivate {
     if (!isSalesman(role)) return false;
 
     const user: UserEntity | undefined | null =
-      await this.authService.findByEmail(result['email']);
+      await this.userService.findByEmail(result['email']);
     // Checking if the user exists and isActive
     if (user && !user.isActive) return false;
